+++ RTWServer/Game/GameClient.cs
@@ -1,6 +1,7 @@
 using System.Net.Sockets;
 using RTWServer.Enum;
 using RTWServer.ServerCore;
+using RTWServer.ServerCore.Interface;
 
 namespace RTWServer.Game;
