+++ RTWServer/Packet/EchoPacket.cs
@@ -1,5 +1,6 @@
 using RTWServer.Enum;
 using RTWServer.ServerCore;
+using RTWServer.ServerCore.Interface;
 
 namespace RTWServer.Packet;
