+++ RTWServer/Program.cs
@@ -1,7 +1,7 @@
 ﻿using System.Net;
 using Microsoft.Extensions.Logging;
 using RTWServer.Game;
-using RTWServer.ServerCore;
+using RTWServer.ServerCore.implementation;
 
 // TODO : 설정 파일에서 IP 주소와 포트 번호를 읽어와서 사용하도록 수정
 string ipAddress = "127.0.0.1";
