+++ RTWServer/ServerCore/Interface/IPacketHandler.cs
@@ -1,5 +1,4 @@
-
-namespace RTWServer.ServerCore;
+namespace RTWServer.ServerCore.Interface;
 
 public interface IPacketHandler
 {
